Significantly improved UI/UX Optimized network calls and internet detection for GPRS/EDGE Optimized database use Laid the basis for writing tests on Kotlin Splitted forecast for current and daily Renewed POJO classes for DarkSky and OpenWeatherMap Renamed MapView to MapaView Replaced HomeView to ForecastView Removed AccuWeather, WorldWeatherOnline, YahooWeather

+++ app/src/main/java/ua/in/khol/oleh/touristweathercomparer/model/weather/owm/pojo/hourly/Coord.java
@@ -8,24 +8,24 @@ public class Coord {
 
     @SerializedName("lat")
     @Expose
-    private Double lat;
+    private float lat;
     @SerializedName("lon")
     @Expose
-    private Double lon;
+    private float lon;
 
-    public Double getLat() {
+    public float getLat() {
         return lat;
     }
 
-    public void setLat(Double lat) {
+    public void setLat(float lat) {
         this.lat = lat;
     }
 
-    public Double getLon() {
+    public float getLon() {
         return lon;
     }
 
-    public void setLon(Double lon) {
+    public void setLon(float lon) {
         this.lon = lon;
     }
